Move ContentHeader and Title from /components/content to /components
