+++ src/components/Title/EditTitle/style.less
@@ -1,4 +1,4 @@
-@import '../../../../variables.less';
+@import '../../../variables.less';
 
 .custom-edit-input-icon {
   margin-left: 0.5em;
